move logic to separate file
